The startup.cs class was purified by writing the extensions method in the field where services are registered in the startup.cs class.
